made it so that the + button adds vis component and each corresponding x button removes the vis
the grid goes out of view if there are too many vis (issue 7). the + button goes out of view after scrolling horizontally when reaching the end of the right side of the inital viewable screen (potentially not a problem after issue 7 is resolved). the grid buttons do not affect the anything  as of right now, which should be fine after issues 6 & 8 are resolved.