fixed the issue where only the last svg was being removed when removing a Vis component
To do this, how the information regarding the outcomes was kept was changed in multiverseMatrix.js. Previously, there was a this.outcomeVars and a this.outcomeData, but that has been condensed to a single this.outcomes with elements of {var: ..., data:..., id:...}. The id is needed to make sure the correct svg is removed. Previously, the iterator variable of the #each was attempted to be used as a unique identifier, but that does not work. The way a unique id is created can be seen in addVis.